% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_design_group_sequential.R
\name{deprecated_getDesignCharacteristics}
\alias{deprecated_getDesignCharacteristics}
\title{Get Design Characteristics}
\usage{
deprecated_getDesignCharacteristics(design)
}
\value{
Returns a \code{\link{TrialDesignCharacteristics}} object.
The following generics (R generic functions) are available for this result object:
\itemize{
\item \code{\link[=names.FieldSet]{names}} to obtain the field names,
\item \code{\link[=print.FieldSet]{print}} to print the object,
\item \code{\link[=summary.ParameterSet]{summary}} to display a summary of the object,
\item \code{\link[=plot.ParameterSet]{plot}} to plot the object,
\item \code{\link[=as.data.frame.TrialDesignCharacteristics]{as.data.frame}} to coerce the object to a \code{\link[base]{data.frame}},
\item \code{\link[=as.matrix.FieldSet]{as.matrix}} to coerce the object to a \code{\link[base]{matrix}}.
}
}
\description{
Calculates the characteristics of a design and returns it.
}
\details{
Calculates the inflation factor (IF),
the expected reduction in sample size under H1, under H0, and under a value in between H0 and H1.
Furthermore, absolute information values are calculated
under the prototype case testing H0: mu = 0 against H1: mu = 1.
}
